requirements.txt
# =============================================================================
# requirements.txt — All project dependencies with pinned versions
# =============================================================================
# WHY PIN VERSIONS?
#   Unpinned packages (e.g. just "scikit-learn") will install the latest version,
#   which may break your code 6 months later. Pinning ensures that anyone who
#   clones this repo gets the exact same environment you used — this is
#   a core principle of MLOps reproducibility.
#
# INSTALL WITH:
#   pip install -r requirements.txt
# =============================================================================

# Core ML
scikit-learn==1.4.2     # Decision Tree, train/test split, metrics
pandas==2.2.2           # Data loading and manipulation
numpy==1.26.4           # Numerical operations

# MLOps tools
mlflow==2.13.0          # Experiment tracking + model registry
dvc==3.51.2             # Data and pipeline versioning
pathspec==0.11.2  # Required by DVC for .dvcignore parsing

# Data download
kaggle==1.6.14          # Optional: to download Titanic dataset via Kaggle API

# Utilities
pyyaml==6.0.1           # To read params.yaml inside Python scripts